// apps/web/src/entities/product/model/product.ts
export interface ExtraProduct {
    /** 추가상품의 결제수단정보 */
    paymentMeans: PayType;
    /** 추가상품의 가격관련 정보 */
    price: {
        /** 추가상품의 판매가 */
        salePrice: number;
        /** 추가상품의 즉시할인정보 */
        immediateDiscountInfo: {
            /** 추가상품의 즉시할인금액 */
            discountAmt: number;
            /** 추가상품의 즉시할인 시작시간 */
            startDateTime: string;
            /** 추가상품의 즉시할인 종료시간 */
            endDateTime: string;
        };
        /** 추가상품의 쿠폰할인정보 */
        couponDiscountInfo: {
            /** 추가상품의 최대 쿠폰할인금액 */
            maxCouponAmt: number;
            /** 추가상품의 쿠폰할인금액 */
            couponDiscountAmt: number;
        };
        /** 추가상품의 추가할인정보 */
        additionalDiscountInfo: {
            /** 추가상품의 추가할인금액 */
            discountAmt: number;
            /** 추가상품의 추가할인 시작일 */
            startDateTime: string;
            /** 추가상품의 추가할인 종료일 */
            endDateTime: string;
        };
    };
    /** 추가상품의 미성년자구매가능여부 */
    minorPurchaseYn: string;
    /** 추가상품의 이미지 */
    imageUrl: string;
    /** 추가상품의 옵션정보 */
    optionInfo: {
        /** 추가상품의 옵션타입 */
        optionType: optionType;
        /** 분리형 옵션 */
        multiOptions: {
            /** 필수 옵션 여부 */
            isRequiredOption: boolean;
            children: {
                /** 할인적용가 */
                buyPrice: number;
                /** 필수 옵션 여부 */
                isRequiredOption: boolean;
                /** 예약재고수량 */
                reservationStockCnt: number;
                /** 옵션판매타입(AVAILABLE, SOLD_OUT) */
                saleType: string;
                /** 옵션값 */
                optionValue: string;
                /** 추가금액 */
                addPrice: number;
                /** 재고수량 */
                stockCnt: number;
                /** 옵션번호 */
                optionNo: number;
                /** 옵션명 */
                optionName: string;
                /** 품절노출 세팅 여부 */
                forcedSoldOut: boolean;
            }[];
            /** 옵션값 */
            optionValue: string;
            /** 옵션명 */
            optionName: string;
        }[];
        /** 일체형 옵션 */
        flatOptions: {
            /** 필수 옵션 여부 */
            isRequiredOption: boolean;
            /** 할인적용가 */
            buyPrice: number;
            /** 예약재고수량 */
            reservationStockCnt: number;
            /** 옵션판매타입 */
            saleType: string;
            /** 옵션값 */
            optionValue: string;
            /** 추가금액 */
            addPrice: number;
            /** 재고수량 */
            stockCnt: number;
            /** 옵션번호 */
            optionNo: number;
            /** 옵션명 */
            optionName: string;
            /** 품절노출 세팅 여부 */
            forcedSoldOut: boolean;
        }[];
        /** 구매자 작성형 정보(텍스트 옵션 내 기입문장) */
        inputs: {
            /** 매칭타입 */
            inputMatchingType: InputMatchingType;
            /** 텍스트 옵션 입력 문구 */
            inputLabel: string;
            /** 필수 여부 */
            required: boolean;
            /** 텍스트 옵션 번호 */
            inputNo: number;
        }[];
        /** 추가상품의 선택옵션타입 */
        optionSelectType: OptionSelectType;
    };
    /** 재고노출여부 */
    displayableStock: boolean;
    /** 추가상품의 상품명 */
    productName: string;
    /** 추가상품의 상품번호 */
    productNo: number;
    /** 추가상품의 구매제한 */
    limitations: {
        /** 추가상품의 환불불가 타입 */
        nonRefundTypes: (boolean | string | number)[];
        /** 추가상품의 최소구매수량 */
        minBuyCnt: number;
        /** 추가상품의 1회당 최대 구매 수량 */
        maxBuyTimeCnt: number;
        /** 추가상품의 최대구매수량 기간 제한 - 기간 */
        maxBuyDays: number;
        /** 추가상품의 최대구매수량 기간 제한 - 제한 수량 */
        maxBuyPeriodCnt: number;
        /** 추가상품의 인당 최대구매수량 */
        maxBuyPersonCnt: number;
        /** 추가상품의 환불가능여부 */
        refundable: boolean;
        /** 추가상품의 네이버페이 가능여부 */
        naverPayHandling: boolean;
        /** 추가상품의 회원전용여부 */
        memberOnly: boolean;
        /** 추가상품의 장바구니 가능여부 */
        canAddToCart: boolean;
    };
}